update : 전체 DTO 정규표현식 업데이트

+++ server/src/main/java/com/codemouse/salog/ledger/budget/dto/BudgetDto.java
@@ -4,6 +4,8 @@ import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.Setter;
 
+import javax.validation.constraints.Max;
+import javax.validation.constraints.Min;
 import java.time.LocalDate;
 
 public class BudgetDto {
@@ -11,14 +13,14 @@ public class BudgetDto {
     @Getter
     public static class Post {
         private LocalDate date;
-        private long budget;
+        private int budget;
     }
 
     @AllArgsConstructor
     @Getter
     public static class Patch {
         private LocalDate date;
-        private long budget;
+        private int budget;
     }
 
     @AllArgsConstructor
@@ -27,7 +29,7 @@ public class BudgetDto {
     public static class Response {
         private long budgetId;
         private LocalDate date;
-        private long budget;
+        private int budget;
         private long totalOutgo; // 지출 월별 합계 계산 (비교를 위함)
         private int dayRemain;
     }
